+++ pdfio-file.c
@@ -1210,6 +1210,8 @@ pdfioFileSetPermissions(
   if (encryption == PDFIO_ENCRYPTION_NONE)
     return (true);
 
+  pdf->encrypt_metadata = true;
+
   return (_pdfioCryptoLock(pdf, permissions, encryption, owner_password, user_password));
 }