+++ CHANGES.md
@@ -12,7 +12,8 @@ v1.6.0 - YYYY-MM-DD
 v1.5.3 - YYYY-MM-DD
 -------------------
 
-- Fixed decryption of PDF files "protected" by 40-bit RC4.
+- Fixed decryption of PDF files "protected" by 40-bit RC4 (Issue #42)
+- Fixed decryption of UTF-16 strings (Issue #42)
 
 
 v1.5.2 - 2025-04-12