+++ CHANGES.md
@@ -8,6 +8,7 @@ v1.5.3 - YYYY-MM-DD
 - Fixed decryption of PDF files "protected" by 40-bit RC4 (Issue #42)
 - Fixed decryption of UTF-16 strings (Issue #42)
 - Fixed decryption of PDF files with large permission values.
+- Fixed support for EncryptMetadata key in the encryption dictionary.
 
 
 v1.5.2 - 2025-04-12